SK_Shieldus_Rookies/wifi-safety-windows-v14/requirements.txt
matplotlib
numpy
openai
pandas
pillow
python-dotenv
reportlab
scipy
ttkbootstrap